composer install faster and without progress bars
We don't need the git clones inside the container. Install with
"--prefer-dist" is faster, use that as a new default. And we don't need
the progress bars inside the container logs, so "--no-progress".

Do not suggest changing that in the docker-compose.yml, the defaults are
most of the times good enough.